+++ src/components/admin/games/index.ts
@@ -1,0 +1,4 @@
+export { GameContentModal } from "./GameContentModal";
+export { HangmanContentEditor } from "./HangmanContentEditor";
+export { WordSearchContentEditor } from "./WordSearchContentEditor";
+export { MemoryContentEditor } from "./MemoryContentEditor";